# contracts/openapi/core-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Katha core-api",
    "description": "Public mobile + web API. Money is an append-only ledger; the client computes no prices or entitlements.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/progress": {
      "put": {
        "tags": [
          "engagement"
        ],
        "summary": "Put Progress",
        "description": "Batch upload of watch progress (the client flushes periodically).",
        "operationId": "put_progress_v1_progress_put",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressBatchBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContinueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/continue": {
      "get": {
        "tags": [
          "engagement"
        ],
        "summary": "Get Continue",
        "description": "Continue watching, most recent first. `limit` sizes the rail (20) or the\nfull-list screen (up to 100).",
        "operationId": "get_continue_v1_me_continue_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContinueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/list": {
      "get": {
        "tags": [
          "engagement"
        ],
        "summary": "Get List",
        "operationId": "get_list_v1_me_list_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/list/{slug}": {
      "put": {
        "tags": [
          "engagement"
        ],
        "summary": "Add List",
        "operationId": "add_list_v1_me_list__slug__put",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "engagement"
        ],
        "summary": "Remove List",
        "operationId": "remove_list_v1_me_list__slug__delete",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/reminders": {
      "get": {
        "tags": [
          "engagement"
        ],
        "summary": "Get Reminders",
        "operationId": "get_reminders_v1_me_reminders_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemindersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/reminders/{slug}": {
      "put": {
        "tags": [
          "engagement"
        ],
        "summary": "Add Reminder",
        "operationId": "add_reminder_v1_me_reminders__slug__put",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemindersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "engagement"
        ],
        "summary": "Remove Reminder",
        "operationId": "remove_reminder_v1_me_reminders__slug__delete",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemindersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/rewards/checkin": {
      "post": {
        "tags": [
          "engagement"
        ],
        "summary": "Checkin",
        "description": "Daily check-in: +5 bonus coins, idempotent per day via the ledger key.",
        "operationId": "checkin_v1_rewards_checkin_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckinResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/push/register": {
      "post": {
        "tags": [
          "engagement"
        ],
        "summary": "Push Register",
        "description": "APNs device-token registration \u2014 the server can then reach this\ndevice for episode drops. Tokens are per (user, device), idempotent.",
        "operationId": "push_register_v1_push_register_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Push Register V1 Push Register Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/invoices": {
      "get": {
        "tags": [
          "engagement"
        ],
        "summary": "My Invoices",
        "description": "Invoices for this account's WEB purchases (buy on web, see in-app).",
        "operationId": "my_invoices_v1_me_invoices_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response My Invoices V1 Me Invoices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/grievance": {
      "post": {
        "tags": [
          "engagement"
        ],
        "summary": "File Grievance",
        "operationId": "file_grievance_v1_grievance_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrievanceIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response File Grievance V1 Grievance Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinResponse": {
        "properties": {
          "granted_coins": {
            "type": "integer",
            "title": "Granted Coins"
          },
          "already_claimed": {
            "type": "boolean",
            "title": "Already Claimed"
          },
          "day": {
            "type": "string",
            "title": "Day"
          },
          "wallet": {
            "$ref": "#/components/schemas/WalletResponse"
          }
        },
        "type": "object",
        "required": [
          "granted_coins",
          "already_claimed",
          "day",
          "wallet"
        ],
        "title": "CheckinResponse"
      },
...
      "ContinueItem": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "episode_id": {
            "type": "string",
            "title": "Episode Id"
          },
          "position_ms": {
            "type": "integer",
            "title": "Position Ms"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "percent": {
            "type": "integer",
            "title": "Percent"
          },
          "series_title": {
            "type": "string",
            "title": "Series Title",
            "default": ""
          },
          "episode_title": {
            "type": "string",
            "title": "Episode Title",
            "default": ""
          },
          "cover_url": {
            "type": "string",
            "title": "Cover Url",
            "default": ""
          },
          "cover_wide_url": {
            "type": "string",
            "title": "Cover Wide Url",
            "default": ""
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "slug",
          "number",
          "episode_id",
          "position_ms",
          "duration_ms",
          "title",
          "percent"
        ],
        "title": "ContinueItem"
      },
      "ContinueResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContinueItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ContinueResponse"
      },
...
      "GrievanceIn": {
        "properties": {
          "contact": {
            "type": "string",
            "title": "Contact"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": "app"
          }
        },
        "type": "object",
        "required": [
          "contact",
          "subject"
        ],
        "title": "GrievanceIn"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MyListResponse": {
        "properties": {
          "slugs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Slugs"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/SeriesSummary"
            },
            "type": "array",
            "title": "Series"
          }
        },
        "type": "object",
        "required": [
          "slugs",
          "series"
        ],
        "title": "MyListResponse"
      },
...
      "ProgressBatchBody": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProgressItemBody"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ProgressBatchBody"
      },
      "ProgressItemBody": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "position_ms": {
            "type": "integer",
            "title": "Position Ms",
            "default": 0
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "default": 0
          },
          "rewind": {
            "type": "boolean",
            "title": "Rewind",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "slug",
          "number"
        ],
        "title": "ProgressItemBody"
      },
      "RemindersResponse": {
        "properties": {
          "slugs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Slugs"
          }
        },
        "type": "object",
        "required": [
          "slugs"
        ],
        "title": "RemindersResponse"
      },
...
      "SeriesSummary": {
        "properties": {
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "episode_count": {
            "type": "integer",
            "title": "Episode Count"
          },
          "primary_language": {
            "type": "string",
            "title": "Primary Language"
          },
          "content_rating": {
            "type": "string",
            "title": "Content Rating",
            "default": "U/A 16+"
          },
          "cover_url": {
            "type": "string",
            "title": "Cover Url",
            "default": ""
          },
          "cover_wide_url": {
            "type": "string",
            "title": "Cover Wide Url",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "slug",
          "title",
          "genres",
          "episode_count",
          "primary_language"
        ],
        "title": "SeriesSummary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WalletResponse": {
        "properties": {
          "balance_bought": {
            "type": "integer",
            "title": "Balance Bought"
          },
          "balance_bonus": {
            "type": "integer",
            "title": "Balance Bonus"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "balance_bought",
          "balance_bonus",
          "total"
        ],
        "title": "WalletResponse"
      },
...
    }
  }
}